+++ frontend/src/composables/userSettingsComposable.ts
@@ -4,8 +4,11 @@ interface UserSettings {
   queueType: number
   sortField?: string
   sortOrder?: number
+  queueSearch: string
+  topicSearch: string
   refetchInterval: number
   darkMode: boolean
+  showGraph: boolean
 }
 
 const storageKey = 'userSettings'
@@ -14,8 +17,11 @@ const settings = reactive<UserSettings>({
   queueType: 1,
   sortField: undefined,
   sortOrder: undefined,
+  queueSearch: '',
+  topicSearch: '',
   refetchInterval: 5000,
   darkMode: false,
+  showGraph: true,
 })
 
 const init = () => {
@@ -37,6 +43,10 @@ const toggleDarkMode = () => {
   loadDarkMode()
 }
 
+const toggleGraph = () => {
+  updateSettings({ ...settings, showGraph: !settings.showGraph })
+}
+
 const loadDarkMode = () => {
   const htmlElement = document.documentElement
 
@@ -52,6 +62,7 @@ export function useUserSettings() {
     settings,
     init,
     toggleDarkMode,
+    toggleGraph,
     updateSettings,
   }
 }