+++ frontend/src/dtos/queue/queueMetricDto.ts
@@ -1,0 +1,9 @@
+export interface QueueMetricDto {
+  queueMetricId: number
+  startTime: string
+  timeSpan: string
+  queueId: number
+  consumeCount: number
+  errorCount: number
+  deadLetterCount: number
+}