+++ frontend/src/enums/queueTypeEnum.ts
@@ -1,0 +1,5 @@
+export enum QueueTypeEnum {
+  PRIMARY = 1,
+  ERROR = 2,
+  DEAD_LETTER = 3,
+}