+++ frontend/src/layouts/AppLayout.vue
@@ -6,15 +6,17 @@ import { useUserSettings } from '@/composables/userSettingsComposable'
 import Listbox from 'primevue/listbox'
 import type { MenuItem } from 'primevue/menuitem'
 import { computed, ref } from 'vue'
-import { useRoute } from 'vue-router'
+import { useRoute, useRouter } from 'vue-router'
+import Graph from './Graph.vue'
 
 const route = useRoute()
+const router = useRouter()
 
 const { isSuccess, isPending, error } = useAuthQuery()
 
 const capitalize = (value: string = '') => value.replace(/\b\w/g, (char) => char.toUpperCase())
 
-const { settings, updateSettings, toggleDarkMode } = useUserSettings()
+const { settings, updateSettings, toggleDarkMode, toggleGraph } = useUserSettings()
 
 const autoRefreshPopover = ref()
 const refetchIntervalOptions = [
@@ -23,31 +25,31 @@ const refetchIntervalOptions = [
     value: 0,
   },
   {
-    label: '1 second',
+    label: '1s',
     value: 1000,
   },
   {
-    label: '5 seconds',
+    label: '5s',
     value: 1000 * 5,
   },
   {
-    label: '30 seconds',
+    label: '30s',
     value: 1000 * 30,
   },
   {
-    label: '1 minute',
+    label: '1m',
     value: 1000 * 60,
   },
   {
-    label: '5 minutes',
+    label: '5m',
     value: 1000 * 60 * 5,
   },
   {
-    label: '30 minutes',
+    label: '30m',
     value: 1000 * 60 * 30,
   },
   {
-    label: '1 hour',
+    label: '1h',
     value: 1000 * 60 * 60,
   },
 ]
@@ -63,6 +65,9 @@ const items = computed((): MenuItem[] => {
   if (route.name === 'messages') {
     items.push({
       label: 'Queues',
+      command: () => {
+        router.push({ name: 'queues' })
+      },
     })
 
     items.push({
@@ -78,49 +83,65 @@ const items = computed((): MenuItem[] => {
 })
 
 const autoRefreshLabel = computed(() => {
-  return `Auto-Refresh (${refetchIntervalOptions.find((x) => x.value === settings.refetchInterval)?.label})`
+  return `${refetchIntervalOptions.find((x) => x.value === settings.refetchInterval)?.label}`
 })
 </script>
 
 <template>
-  <div v-if="!isPending && isSuccess" class="flex h-screen flex-col">
-    <div class="flex items-center border-b px-4 pb-4 pt-4 dark:border-b-surface-700">
-      <div class="flex">
-        <div class="flex h-14 w-14 items-center justify-center rounded-xl text-2xl">
-          <img :src="mtLogoUrl" class="w-full dark:hidden" />
-          <img :src="mtLogoUrlDark" class="hidden w-full dark:block" />
+  <div v-if="!isPending && isSuccess" class="flex h-screen w-full flex-col">
+    <Popover ref="autoRefreshPopover">
+      <div class="flex w-72 flex-col gap-2">
+        <div>
+          Select an interval to automatically refresh the queues and messages view. We plan to
+          integrate a real-time, socket-based system for instant updates in a future release.
         </div>
-
-        <div class="ms-4 flex flex-col justify-center">
-          <div class="text-2xl font-semibold text-primary">MassTransit</div>
-          <div class="flex items-center gap-2">
-            <Breadcrumb style="padding: 0" :model="items" />
-          </div>
-        </div>
+        <Listbox
+          :options="refetchIntervalOptions"
+          :model-value="settings.refetchInterval"
+          @update:model-value="onRefreshIntervalChange"
+          option-value="value"
+          option-label="label"
+        ></Listbox>
       </div>
-      <div class="my-auto me-3 ms-auto items-center">
-        <Button
-          @click="(e) => autoRefreshPopover.toggle(e)"
-          :label="autoRefreshLabel"
-          text
-        ></Button>
-        <Popover ref="autoRefreshPopover">
-          <div class="flex w-72 flex-col gap-2">
-            <div>
-              Select an interval to automatically refresh the queues and messages view. We plan to
-              integrate a real-time, socket-based system for instant updates in a future release.
-            </div>
-            <Listbox
-              :options="refetchIntervalOptions"
-              :model-value="settings.refetchInterval"
-              @update:model-value="onRefreshIntervalChange"
-              option-value="value"
-              option-label="label"
-            ></Listbox>
-          </div>
-        </Popover>
+    </Popover>
 
-        <Button @click="toggleDarkMode" icon="pi pi-palette" text class="ms-1"></Button>
+    <div class="flex">
+      <div class="flex grow flex-col">
+        <div class="flex items-center border-b px-4 pb-4 pt-4 dark:border-b-surface-700">
+          <div class="flex">
+            <div class="flex h-14 w-14 items-center justify-center rounded-xl text-2xl">
+              <img :src="mtLogoUrl" class="w-full dark:hidden" />
+              <img :src="mtLogoUrlDark" class="hidden w-full dark:block" />
+            </div>
+
+            <div class="ms-4 flex flex-col justify-center">
+              <div class="text-2xl font-semibold text-primary">MassTransit</div>
+              <div class="flex items-center gap-2">
+                <Breadcrumb style="padding: 0" :model="items" />
+              </div>
+            </div>
+          </div>
+
+          <div class="my-auto me-3 ms-auto items-center">
+            <Button @click="toggleDarkMode" icon="pi pi-palette" text class="ms-1"></Button>
+            <Button
+              @click="(e) => autoRefreshPopover.toggle(e)"
+              :label="autoRefreshLabel"
+              icon="pi pi-hourglass "
+              text
+            ></Button>
+            <Button
+              @click="toggleGraph"
+              :icon="`pi pi-angle-${settings.showGraph ? 'right' : 'left'}`"
+              text
+              class="ms-1"
+            ></Button>
+          </div>
+        </div>
+        <slot name="menu"></slot>
+      </div>
+      <div class="flex">
+        <slot name="right"></slot>
       </div>
     </div>