+++ frontend/src/pages/broker/queues/BrokerQueues.vue
@@ -4,7 +4,10 @@ import { useUserSettings } from '@/composables/userSettingsComposable'
 import { FilterMatchMode } from '@primevue/core/api'
 import Column from 'primevue/column'
 import DataTable, { type DataTableSortEvent } from 'primevue/datatable'
-import { computed, ref } from 'vue'
+import IconField from 'primevue/iconfield'
+import InputIcon from 'primevue/inputicon'
+import InputText from 'primevue/inputtext'
+import { computed, ref, watchEffect } from 'vue'
 import { useRouter } from 'vue-router'
 
 const router = useRouter()
@@ -23,8 +26,17 @@ const selectQueue = (data: any) => {
   })
 }
 
+const search = ref(settings.queueSearch)
+
+watchEffect(() => {
+  updateSettings({
+    ...settings,
+    queueSearch: search.value,
+  })
+})
+
 const filters = ref({
-  queueName: { value: null, matchMode: FilterMatchMode.CONTAINS },
+  queueName: { value: search, matchMode: FilterMatchMode.CONTAINS },
 })
 
 const onSort = (e: DataTableSortEvent) => {
@@ -37,14 +49,15 @@ const onSort = (e: DataTableSortEvent) => {
 </script>
 
 <template>
-  <div>
+  <div class="">
     <DataTable
+      show-gridlines
       scrollable
       data-key="queueName"
       scroll-height="flex"
       :value="queuesView"
       removable-sort
-      class="grow overflow-auto"
+      class="rq-grid grow overflow-auto"
       selection-mode="single"
       striped-rows
       v-model:filters="filters"
@@ -54,19 +67,36 @@ const onSort = (e: DataTableSortEvent) => {
       @row-select="(e) => selectQueue(e.data)"
       @sort="onSort"
     >
-      <Column sortable field="queueName" header="Name" class="overflow-hidden overflow-ellipsis">
-        <template #body="{ data }">
-          {{ data.queueName }}
-        </template>
-        <template #filter="{ filterModel, filterCallback }">
-          <InputText
-            v-model="filterModel.value"
-            type="text"
-            @input="filterCallback()"
-            placeholder="Search by queue name"
-          />
+      <Column field="queueName" class="overflow-hidden overflow-ellipsis py-0">
+        <template #header>
+          <div class="flex w-full items-center">
+            <b>Name</b>
+            <IconField class="ms-1 grow">
+              <InputIcon class="pi pi-search" v-if="!search" />
+              <InputIcon class="pi pi-times cursor-pointer" v-else @click="search = ''" />
+              <InputText
+                placeholder="Search anything..."
+                class="w-full border-0 shadow-none dark:bg-surface-900"
+                v-model="search"
+                ref="searchInputText"
+              />
+            </IconField>
+          </div>
         </template>
       </Column>
+      <Column field="queueAutoDelete" header="Auto Delete" class="w-[0] whitespace-nowrap">
+        <template #body="{ data }">
+          <div v-show="data['queueAutoDelete']" class="text-nowrap">
+            {{ data['queueAutoDelete'] / 60 }}m
+          </div>
+          <div v-show="!data['queueAutoDelete']">-</div>
+        </template>
+      </Column>
+      <Column
+        field="queueMaxDeliveryCount"
+        header="Max Delivery"
+        class="w-0 whitespace-nowrap"
+      ></Column>
       <Column sortable field="ready" header="Ready" class="w-[0]"> </Column>
       <Column sortable field="errored" header="Errored" class="w-[0]"></Column>
       <Column
@@ -77,14 +107,6 @@ const onSort = (e: DataTableSortEvent) => {
       ></Column>
       <Column sortable field="scheduled" header="Scheduled" class="w-[0]"></Column>
       <Column sortable field="locked" header="Locked" class="w-[0]"></Column>
-      <Column sortable field="queueAutoDelete" header="Auto Delete" class="w-[0] whitespace-nowrap">
-        <template #body="{ data }">
-          <div v-show="data['queueAutoDelete']" class="text-nowrap">
-            {{ data['queueAutoDelete'] / 60 }}m
-          </div>
-          <div v-show="!data['queueAutoDelete']">-</div>
-        </template>
-      </Column>
     </DataTable>
   </div>
 </template>